requirements.txt
Django==5.0.2
environs[django]==10.3.0
djangorestframework==3.14.0
factory-boy==3.3.0
psycopg2-binary==2.9.9